# add buildHTML helper that accepts *strings.Builder