--- a/Makefile
+++ b/Makefile
@@ -3,3 +3,6 @@ test:
 
 run:
 	go run examples/main.go
+
+bench:
+	go test -bench .