--- a/dom_test.go
+++ b/dom_test.go
@@ -4,20 +4,22 @@ import (
 	"bytes"
 	"fmt"
 	"html/template"
+	"strings"
 	"testing"
 
 	"github.com/choonkeat/dom-go"
 )
 
 func TestAttribute(t *testing.T) {
-	got := dom.Div(
+	var gotHTML, wantHTML strings.Builder
+	dom.Div(
 		dom.Attrs(
 			"class", "greeting",
 			"style", "color: red;",
 		),
 		dom.Text("Hello, world!"),
-	)
-	want := dom.Node{
+	).HTML(&gotHTML)
+	dom.Node{
 		Name: "div",
 		Attributes: []dom.Attribute{
 			{Name: "class", ValueText: "greeting"},
@@ -26,9 +28,9 @@ func TestAttribute(t *testing.T) {
 		Children: []dom.Node{
 			{InnerText: "Hello, world!"},
 		},
-	}
-	if got.HTML() != want.HTML() {
-		t.Fatalf("want %#v but got %#v", want.HTML(), got.HTML())
+	}.HTML(&wantHTML)
+	if gotHTML.String() != wantHTML.String() {
+		t.Fatalf("want %#v but got %#v", wantHTML.String(), gotHTML.String())
 	}
 }
 
@@ -42,7 +44,7 @@ func Example() {
 				dom.Attrs(),
 				dom.Text("10"),
 			),
-		).HTML(),
+		).HTML(nil).String(),
 	)
 	// Output: <div class="1 2 3" data-foo="4&lt;&#39;&#34;5&#34;&#39;&gt;6">&lt;oops&gt;789&lt;/oops&gt;<strong>10</strong></div>
 }
@@ -54,7 +56,7 @@ func ExampleAttrs() {
 				"href", "https://google.com",
 				"target", "_blank",
 			),
-		).HTML(),
+		).HTML(nil).String(),
 	)
 	// Output: <div href="https://google.com" target="_blank"></div>
 }
@@ -71,24 +73,24 @@ func ExampleElement() {
 				dom.Attrs(),
 				dom.Text("Google"),
 			),
-		).HTML(),
+		).HTML(nil).String(),
 	)
 	// Output: <a href="https://google.com" target="_blank">Goo&lt;g&gt;le<blockquote>Google</blockquote></a>
 }
 
 func ExampleText() {
 	fmt.Println(
-		dom.Text("Goo<g>le").HTML(),
+		dom.Text("Goo<g>le").HTML(nil).String(),
 	)
 	// Output: Goo&lt;g&gt;le
 }
 
 func BenchmarkAttrHTML(b *testing.B) {
-
 	for i := 0; i < b.N; i++ {
+		var sb strings.Builder
 		dom.Attrs(
 			"href", "https://google.com",
-		)[0].HTML()
+		)[0].HTML(&sb).String()
 	}
 	b.ReportAllocs()
 	b.ReportMetric(float64(b.N), "AttrHTML")
@@ -96,6 +98,7 @@ func BenchmarkAttrHTML(b *testing.B) {
 
 func BenchmarkNodeHTML(b *testing.B) {
 	for i := 0; i < b.N; i++ {
+		var sb strings.Builder
 		dom.A(
 			dom.Attrs(),
 			dom.Text("Goo<g>le"),
@@ -103,7 +106,7 @@ func BenchmarkNodeHTML(b *testing.B) {
 				dom.Attrs(),
 				dom.Text("Google"),
 			),
-		).HTML()
+		).HTML(&sb).String()
 	}
 	b.ReportAllocs()
 	b.ReportMetric(float64(b.N), "NodeHTML")