worth it?

[before](https://github.com/choonkeat/dom-go/actions/runs/6779812005/job/18427465970)
```

BenchmarkAttrHTML-2       	 5737236	       211.7 ns/op	   5737236 AttrHTML	      80 B/op	       2 allocs/op
BenchmarkNodeHTML-2       	 1534218	       786.0 ns/op	   1534218 NodeHTML	     232 B/op	       9 allocs/op
BenchmarkHtmlTemplate-2   	  143379	      8083 ns/op	    143379 HTMLTemplate	    1384 B/op	      53 allocs/op

```

[after](https://github.com/choonkeat/dom-go/actions/runs/6780041183/job/18428070446)
```

BenchmarkAttrHTML-2       	 4646418	       252.3 ns/op	   4646418 AttrHTML	     128 B/op	       4 allocs/op
BenchmarkNodeHTML-2       	 1818426	       657.4 ns/op	   1818426 NodeHTML	     240 B/op	       8 allocs/op
BenchmarkHtmlTemplate-2   	  140749	      8541 ns/op	    140749 HTMLTemplate	    1384 B/op	      53 allocs/op

```